base64_decode: always cast to unsigned char when using the input characters as indexes in LUTs
char's signedness is implementation-defined, with most implementations
defaulting to signed, so characters above 127 are actually treated as
negative.

For this reason, before this fix if invalid base64 data containing >127
characters was passed to l8w8jwt_base64_decode, it indexed lookup tables
with negative values, reading garbage values from the stack (and
potentially crashing if it bumped on a guard page or an unallocated
stack portion).